move debug related class and stubs to a separate namespace
