+++ tests/Rules/Debug/NoDebugInBladeTest.php
@@ -1,26 +1,26 @@
 <?php declare(strict_types=1);
 
-namespace Rules;
+namespace Rules\Debug;
 
-use Hihaho\PhpstanRules\Rules\NoDebugInBlade;
+use Hihaho\PhpstanRules\Rules\Debug\NoDebugInBladeRule;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use PHPUnit\Framework\Attributes\Test;
 
 /**
- * @extends RuleTestCase<NoDebugInBlade>
+ * @extends RuleTestCase<NoDebugInBladeRule>
  */
 class NoDebugInBladeTest extends RuleTestCase
 {
     protected function getRule(): Rule
     {
-        return new NoDebugInBlade();
+        return new NoDebugInBladeRule();
     }
 
     #[Test]
     public function no_debug_statements_should_be_present_in_blade(): void
     {
-        $this->analyse([__DIR__ . '/stubs/debug-in-view.blade.php'], [
+        $this->analyse([__DIR__ . '/stubs/debug-in-view-stub.blade.php'], [
             [
                 'No debug directives should be present in blade files.',
                 1,
