+++ tests/Rules/Debug/stubs/debug-in-view-stub.blade.php
